Update naming conventions
