+++ DSA-1/LAB BST validity checker/MyBSTChecker.h
@@ -1,0 +1,15 @@
+#ifndef MYBSTCHECKER_H
+#define MYBSTCHECKER_H
+
+// Your code here (include additional header files, if needed)
+#include "MyTreeNode.h"
+
+class MyBSTChecker {
+ public:
+  static MyTreeNode* CheckBSTValidity(MyTreeNode* rootNode) {
+    // Your code here (remove the placeholder line below)
+    return rootNode;
+  }
+};
+
+#endif
