+++ DSA-1/LAB BST validity checker/main.cpp
@@ -1,28 +1,27 @@
 #include <iostream>
 #include <string>
-#include "Node.h"
-#include "BSTChecker.h"
+
+#include "MyBSTChecker.h"
+#include "MyTreeNode.h"
 using namespace std;
 
-int main(int argc, char *argv[]) {
-   // Get user input
-   string userInput;
-   getline(cin, userInput);
-   
-   // Parse into a binary ree
-   Node* userRoot = Node::Parse(userInput);
-   if (userRoot) {
-      Node* badNode = BSTChecker::CheckBSTValidity(userRoot);
-      if (badNode) {
-         cout << to_string(badNode->key) << endl;
-      }
-      else {
-         cout << "No violation" << endl;
-      }
-   }
-   else {
-      cout << "Failed to parse input tree" << endl;
-   }
-    
-   Node::DeleteTree(userRoot);
+int main(int argc, char* argv[]) {
+  // Get user input
+  string userInput;
+  getline(cin, userInput);
+
+  // Parse into a binary ree
+  MyTreeNode* userRoot = MyTreeNode::Parse(userInput);
+  if (userRoot) {
+    MyTreeNode* badNode = MyBSTChecker::CheckBSTValidity(userRoot);
+    if (badNode) {
+      cout << to_string(badNode->key) << endl;
+    } else {
+      cout << "No violation" << endl;
+    }
+  } else {
+    cout << "Failed to parse input tree" << endl;
+  }
+
+  MyTreeNode::DeleteTree(userRoot);
 }
